Grade the active-set QP against Maros-Meszaros ground truth
The active-set engine had almost no benchmark exposure: it was graded
on 3 hand-constructed problems in adversarial_convex.md, while the IPM
path was graded on all 138 Maros-Meszaros instances with published
optima. Both sets were already in the tree; nothing pointed them at
this solver. Suggested by Jeff Renfro.

Adds qp-active-set as a fourth column to the existing runner (renamed
three_way -> four_way, since it now reports four solvers), on the same
.nl files and the same BPMPD ground truth.

Result: 46/138 correct, against qp-ipm's 137/138. Failure modes:

    72  SearchDirectionBecomesTooSmall
    14  TimeOut
     4  InternalError
     1  MaximumIterationsExceeded
     1  ParseError

Two things make this more tractable than the raw number suggests.
One dominant symptom rather than 92 unrelated ones, which reads as a
single stalling/anti-cycling defect. And zero solved-but-wrong: every
failure is loud, where Clarabel has 9 and the NLP path 8 wrong
objectives. It also wins 10 problems outright, including 4 of the
LISWET family where both Clarabel and NLP return wrong objectives.

Verified not caused by the #398/#401/#402/#403 merges: a binary built
at bebfce44 (pre-#398) gives bit-identical status and objective on 8
sampled failures. Long-standing, previously unmeasured.

Harness fixes made along the way, each after being misled by it:

- Per-problem progress now flushes. Redirected to a file, stdout is
  block-buffered, so progress lagged ~55 problems behind reality and
  read exactly like a hang. Cost one killed run.
- Speed-set membership is `solve_time is not None`, not truthiness.
  0.0 is falsy, so the old test dropped every sub-millisecond solve --
  which was *every* correct active-set solve, and it crashed summarize
  via geomean-of-empty returning None.
- The active-set speed section reports that the data is unavailable
  rather than synthesizing one: that path writes
  total_wallclock_time_secs = 0.0 into --json-output while really
  taking 0.568s on AUG2D (qp-ipm reports 0.678s correctly on the same
  problem). A geomean over that field would read "instantaneous".
- Reports now record the producing binary's --about line. The previous
  committed report was six weeks stale and its NLP column had drifted
  129/138 -> 105/138 on binary drift alone, with nothing to reveal it.
